Configure the protobuf include paths
By default repo-config adds this repository (api-common) as a dependency
when generating new API repos, as it is supposed to be included by most
(if not all) APIs. So the default `include_paths` repo-config uses looks
like `["submodules/api-common-protos",
"submodules/frequenz-api-common/proto"].

Of course in this repository we don't have ourself as a cyclic
dependency, and we only depend on `submodules/api-common-protos`, so we
need to override the `include_paths` option, otherwise we get errors
about `submodules/frequenz-api-common/proto` not existing.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>