Add country code to the Location message
Some control parameters change depending upon the country of the microgrid.
While the country code can be derived from the geo-coordinates, it is
better to expose it via our APIs, so that the behaviour of obtaining the
codes is more consistent, and clients do not need to worry about it.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>