Add versioning to common proto files
This commit adds versioning to the common proto files. This is done by
adding a v1 directory to the common directory and moving all the proto
files into it. The proto packages are also renamed to include the v1
suffix.

This commit also updates the imports in the test_common.py file to
reflect the changes.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>