Add new sensor category and metric enums
In our terminology, `components` are the devices that are connected to
a microgrid's electrical circuit, and `sensors` are the devices that
measure the physical parameters of the environment.

Hence, we need to distinguish between the two, and we need to be able
to identify the type of sensor and the type of metric that it measures.

This commit adds the `frequenz.api.common,v1.sensors` package, which
contains the `SensorCategory` and `SensorMetric` enums.

Also, this commit removes
* the `COMPONENT_CATEGORY_SENSOR` variant from the `ComponentCategory`
  enum, and
* The `SensorType` enum.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>