Make lower and upper bounds optional in the Bounds message
The lower and upper bounds in the Bounds message are now optional. If
absent, there is no lower or upper bound. This is useful for metrics
that are intended to have only one bound.

Setting extreme values can be less intuitive to understand when
inspecting the data. Someone looking at the bounds might wonder why they
are set to such extreme values and could interpret them incorrectly.

Making it optional has the following benefits:

Explicitness: Making the fields optional makes it explicit that a bound
was not set, making it easier to understand when reading the data.

Flexibility: Making the bounds optional gives you more flexibility in
dealing with edge cases. For example, if you later decide to handle
undefined bounds differently, you can easily do so.

Semantic Accuracy: Using optional fields for values that are truly
optional is semantically more accurate.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>