Replace inclusion and exclusion bounds with a single array of bounds
This commit removes the `system_inclusion_bounds` and
`system_exclusion_bounds` fields from the `Metric` and
`MetricAggregation` messages. These fields have been replaced with a
single array of `Bounds` in both messages. This simplifies the message
definition, and removes the requirement of clients having to check if a
parameter is _not_ in a given pair of bounds. This also extends the
possibility of having more than 2 pairs bounds for a given metric.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>