Add a new proto message to represent a sampled sensor metric
This commit adds a new proto message to represent a sampled sensor
metric along with its value. This message will be used to represent
the metric samples in the `SensorData` message.

This is different to the existing `MetricSample` message in the
`frequenz.api.common.v1.metrics` package as it does not contain a list
of bounds.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>