Add a message representing components
This commit adds a new message `ComponentCategoryMetadataVariant` which is
a union type for the metadata/description messages for different component
categories. Each component-category-specific metadata/description message
contains the fields required to minimally describe the component category
further, e.g., its type, or its rated max max current. These metadata
messages can be extended in the future.

In addition to the above, this commit also adds a message `Component`, that
describes components present in a microgrid. This message uses
`ComponentCategoryMetadataVariant` to further describe the component
metadata.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>