Add message representing electrical connections between components
This commit adds a new message `ComponentConnection` to represent
electrical connections between two components installed in a microgrid.
The message is not about data flow but rather about the physical
electrical connections between components. Therefore, the IDs for the
source and destination components correspond to the actual setup within
the microgrid.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>