Add a message describing microgrids
This commit adds the following definition:

`Microgrid`: A microgrid is a localized grouping of electricity
generation, energy storage, and loads that normally operates connected
to a traditional centralized grid. Each microgrid has a unique
identifier and is associated with an enterprise account. A key feature
is that it has a physical location and is situated in a delivery area.
Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>