Add a `source` field to `MetricSample`
This commit adds an optional string field that can be used to identify the
source of the metric. This is useful in cases where the same metric can be
obtained from multiple sensors or sources in the same component.

E.g., a hybrid inverter can have  multiple DC strings, each with its own
current and voltage sensors. In such cases, the source field can be used to
differentiate between the different sources.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>