Add warnings to sensor `SensorState`
This commit adds a new field `warnings` to the `SensorState` message in
the `sensors.proto` file. This field is a repeated field of type
`SensorDiagnostic`, which is used to represent warnings for the
microgrid sensor. The `warnings` field is expected to have warnings if
and only if the sensor is reporting warnings. The list will contain
unique members, meaning no warning will exist twice in this list.

This makes the `SensorState` message consistent with respect to
`ElectricalComponentState`.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>