Add enum `ElectricalComponentControlMode`
This enum is used to explicitly define whether an electrical component
is intended to be active (enabled) or inactive (disabled). This
intention directly determines the component's availability to downstream
services and applications.

Note that this enum differs from the `ElectricalComponentStatus` enum,
which is used to report the actual operational state of the component.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>

# Conflicts:
#	RELEASE_NOTES.md
#	proto/frequenz/api/common/v1/microgrid/electrical_components/electrical_components.proto