Rename `metric_sample.proto` to `metrics.proto`
This change renames the `metric_sample.proto` file to `metrics.proto` to
better reflect its content. The file contains multiple metric-related
messages, including `SimpleMetricValue` and `MetricSample`, which are
used to represent metric values and samples, respectively. The new name
`metrics.proto` is more descriptive and aligns with the purpose of the
file.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>

# Conflicts:
#	RELEASE_NOTES.md