Remove `SensorMetricSample` in favour of `MetricSample`
This commit removes the `SensorMetricSample` message in favour of using
`MetricSample` for sensors. The `SensorMetricSample` message was
a subset of the `MetricSample` message.

From a design perspective focused on consistency, flexibility, and
minimizing API surface area while handling variations via optional
fields (which is idiomatic in Protobuf3), there's a strong argument that
`SensorMetricSample` is redundant and potentially confusing, and that
having only `MetricSample` is a cleaner design.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>