Rename `SensorStateCode` variant `ON` to `OK` in sensors.proto
Sensor state `ON` is misleading, as it can mislead users to think that
the API controls the on/off state of the sensor. This is not the case,
as the API only provides information about the sensor's state. The
sensor's state can be `OK` (indicating that the sensor is functioning
correctly) or `ERROR` (indicating that the sensor is in an error state).

This change renames the `SENSOR_STATE_CODE_ON` variant to
`SENSOR_STATE_CODE_OK`, which is more accurate and descriptive.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>