Remove `SensorCategory` enum
This enumeration was not useful and potentially confusing. Sensors can
report different sensor metrics, and they could belong to several of
these categories simultaneously. This defeats the purpose of having
singular categories for sensors. We need to rethink how to categorize
sensors. Until then, having it does not add any value, and therefore it
has been removed.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>

# Conflicts:
#	RELEASE_NOTES.md