Reorganize releases notes on upgrading and new features
This commit reorganizes the release notes to move all "breaking changes"
to the Upgrading section, and to group changes by type (removal, rename,
renumbering and addition) and by protobuf package.

This hopefully will make easier for users to know how to deal with
changes and what new fetures are available, depending on which packages
they use/they are interested in.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>