Change type of `metrics.MetricSample.connection` field
This commit changes the type of the `metrics.MetricSample.connection`
field from `string` to a new message type `metrics.MetricConnection`.

This change allows for better categorization of the connection type,
which can be useful in control and monitoring applications. The new
`MetricConnection` message includes a `MetricConnectionCategory` enum
that categorizes the connection type (e.g., battery, PV array,
temperature`s), and a `name` field that can be used to identify the
specific connection point from which the metric was obtained.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>