Re-create the v1 package with the contents of the v0.6.x branch
This commit re-creates the `v1` package with the contents of the
`v0.6.x` branch, which is the latest stable version of the API. The
previous `v1` package has been renamed to `v1alpha7`.

This means that any client or server code that was using the v0.6.x
branch can now use the main branch without any changes, as the
`v1` package now contains the same contents as the `v0.6.x` branch.

Signed-off-by: Tiyash Basu <tiyash.basu@frequenz.com>