Override the default include path when generating protobuf files
By default `repo-config` will include the
`submodules/frequenz-api-common` directory, which is us, and we don't
want to include ourselves as a submodule, so we need to explicitly set
the `include_paths` to an empty array.

We also need to bump the dependency on `frequenz-repo-config` to
get a bug fix that allows overriding with an empty array.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>